+++ lib/whois/record/scanners/base_shared1.rb
@@ -19,6 +19,7 @@ module Whois
         self.tokenizers += [
             :skip_empty_line,
             :scan_available,
+            :scan_reserved,
             :scan_keyvalue,
         ]
 
@@ -26,6 +27,12 @@ module Whois
         tokenizer :scan_available do
           if @input.skip(/^No Data Found\n/)
             @ast["status:available"] = true
+          end
+        end
+
+        tokenizer :scan_reserved do
+          if settings[:pattern_reserved] && @input.scan(settings[:pattern_reserved])
+            @ast["status:reserved"] = true
           end
         end